Use `/run` for PID only if it exists.
Older systems still use `/var/run` and newer systems have a symbolic link from `/var/run` to `/run`, where their PID files reside. I added a check to see if `/run` exists, since we cannot rely on the symbolic link. This means, `/run` is used if it exists and if not `/var/run` is used.